requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==3.2.2
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
databases==0.9.0
ecdsa==0.19.1
fastapi==0.115.12
h11==0.14.0
httptools==0.6.4
idna==3.10
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
sentence-transformers==2.2.2
qdrant-client==1.7.0
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
huggingface-hub==0.19.4
